Change adapter into vendor string to just be hexadecimal vendor ID.
I realized while reading the WebGPU spec that returning anything other than a "normalized identifier string" there is actually illegal, which the previous value definititely was.

If we want to fill these up nicely, we'll need to keep a database similar to what Dawn is doing. Hell, maybe it's best to literally use the same database for compatibility?